requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cssselect==1.2.0
idna==3.4
jmespath==1.0.1
lxml==4.9.3
numpy==1.25.2
packaging==23.1
pandas==2.1.0
parsel==1.8.1
proxybroker2==2.0.0a4
pyodbc==4.0.39
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
w3lib==2.1.2
